+++ client/src/Components/ContentPage.js
@@ -9,7 +9,8 @@ import MapWithMarkers from "./MapWithMarkers";
 import GrantTellerAccount from "./RoleContent/PointLeaderTransaction/GrantTellerAccount";
 import TransactionStatistic from "./RoleContent/PointLeaderTransaction/TransactionStatistic";
 import UserTracking from "./RoleContent/User/UserTracking";
-import Profile from "./Profile";
+// import Profile from "./Profile";
+import {Profile} from "../Components";
 import LeafletMap from "./LeafletMap";
 import DashboardPage from "./DashboardPage";
 import AuthService from "../services/auth.service";
