--- client/src/Components/Profile.js
@@ -1,29 +1,0 @@
-import React from "react";
-import AuthService from "../services/auth.service";
-
-const Profile = () => {
-  const currentUser = AuthService.getCurrentUser();
-
-  return (
-    <div className="container">
-      <header className="jumbotron">
-        <h3>
-          <strong>{currentUser.username}</strong> Profile
-        </h3>
-      </header>
-      <p>
-        <strong>Id:</strong> {currentUser.id}
-      </p>
-      <p>
-        <strong>Email:</strong> {currentUser.email}
-      </p>
-      <strong>Authorities:</strong>
-      <ul>
-        {currentUser.roles &&
-          currentUser.roles.map((role, index) => <li key={index}>{role}</li>)}
-      </ul>
-    </div>
-  );
-};
-
-export default Profile;
