+++ client/src/Components/Profile/Profile.module.css
@@ -1,0 +1,38 @@
+.container {
+    max-width: 400px;
+    margin: 0 auto;
+    background-color: #ffffff;
+    border-radius: 5px;
+    border: 3px solid teal;
+    padding: 20px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+  
+  .title {
+    text-align: center;
+    font-size: 32px;
+    font-weight: bold;
+    color: teal;
+    letter-spacing: 2px;
+  }
+  
+  .Name {
+    text-align: center;
+    font-size: 32px;
+  }
+  
+  ul {
+    list-style-type: none;
+    padding: 0;
+  }
+  
+  ul li:before {
+    content: "•";
+    color: teal;
+    display: inline-block;
+    width: 1em;
+    margin-left: -1em;
+  }
