+++ client/src/Components/index.js
@@ -1,5 +1,7 @@
 export { default as Header } from './Header';
 export { default as Content } from './Content';
 export { default as SideBarCT } from './SideBarCT';
+export { default as Profile } from './Profile';
+
